More accurate check for prefix argument.
Someone might want to explicitly give a prefix of one. In the previous
implementation, such an event would call the relevant hungry-*-iter,
which *might indeed* delete one forward or backward character, but only
if that character was not whitespace. If it was whitespace, it would get
HUNGRY!

This implementation tests whether or not a prefix has explicitly been
given, not whether or not the prefix is equal to one. Thus, someone can
explicitly use these functions to delete one whitespace character, as
well as any number of anything characters.