Bugfix (see description)
This commit fixes reluctant hungry deletion in the case where the to-be
joined words are seperated by a single whitespace character.

Previously, hungry deletion would behave surprisingly when the separating
character was not a space. For example, pressing backspace from

```
foo
bar
^
```

would produce

```
foobar
   ^
```

instead of

```
foo bar
    ^
```

as it should have